more test reorganisation
