+++ test/nmodl/test_kinetic.py
@@ -11,10 +11,8 @@ from kinetic import cadifusCell
 def test_kinetic():
     cadifus_cell = cadifusCell()
     cadifus_cell.record()
-    coreneuron_enable = bool(
-        strtobool(os.environ.get("NRN_CORENEURON_ENABLE", "false"))
-    )
-    coreneuron_gpu = bool(strtobool(os.environ.get("CORENRN_ENABLE_GPU", "false")))
+    coreneuron_enable = strtobool(os.environ.get("NRN_CORENEURON_ENABLE", "false"))
+    coreneuron_gpu = strtobool(os.environ.get("CORENRN_ENABLE_GPU", "false"))
     cadifus_cell.simulate(1, 0.1, coreneuron_enable, coreneuron_gpu)
     assert np.isclose([cadifus_cell.record_vectors["ca_ion[0]"][2]], [5e-05])
